refactor: move check output contracts to output
